Fixed language server smoke tests
