+++ news/3 Code Health/3684.md
@@ -1,0 +1,1 @@
+Fixed language server smoke tests.
